backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
pydantic>=2.7.0
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
openai>=1.30.0
langchain>=0.2.0
langchain-openai>=0.1.8
python-multipart>=0.0.9
aiofiles>=23.2.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
httpx>=0.27.0
